requirements.txt
streamlit==1.31.1
streamlit-option-menu==0.3.12
pillow==10.2.0
google-generativeai==0.4.0